pyproject.toml
[project]
name = "mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.11.0.post1",
    "httpx>=0.28.1",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-google-firestore>=0.5.0",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.28",
    "langchain-pinecone>=0.2.11",
    "mcp[cli]>=1.12.3",
    "openai>=1.99.1",
    "pinecone>=7.3.0",
    "pymupdf>=1.26.3",
    "pypdf>=5.9.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.0",
]
